Input width as an Integer
Google changed the way they managed the optional width parameter. A string "200" etc used to work fine, but from this morning (in some EU zones) it looks like we need to use an Integer to get the button displayed.

Feel free to double check this information. I also opened a ticket with Google to check this.